+++ GilCat/GilCat/DiaryView.swift
