--- GilCat/GilCat/Resources/Assets.xcassets/Morphic1.colorset/Contents.json
@@ -1,20 +1,0 @@
-{
-  "colors" : [
-    {
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "0.584",
-          "green" : "0.494",
-          "red" : "0.420"
-        }
-      },
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
