+++ GilCat/GilCat/Sources/Global/Extensions/TestExtension.swift
@@ -8,8 +8,8 @@
 import Foundation
 import SwiftUI
 
-extension Color{
-    static let mainBlue = Color("mainBlue")
-    static let mainOrange = Color("mainOrange")
-    static let mainBlack = Color("mainBlack")
+extension Color {
+    static let mainBlue = Color("BackGroundColor")
+    static let mainOrange = Color("ButtonColor")
+    static let mainBlack = Color("PickerColor")
 }
